[ui] add billing view in services and iaas

+++ scripts/database/mysql/schemas/Schemas.sql
@@ -1621,8 +1621,8 @@ CREATE TABLE `billinglimit` (
   `service_provider_id` int(8) unsigned NOT NULL,
   `repeats` int(16) NOT NULL,
   `repeat_day` int(16) NOT NULL,
-  `repeat_start_time` int(16) NOT NULL,
-  `repeat_end_time` int(16) NOT NULL,
+  `repeat_start_time` BIGINT(16) NOT NULL,
+  `repeat_end_time` BIGINT(16) NOT NULL,
   PRIMARY KEY (`id`),
   FOREIGN KEY (`id`) REFERENCES `entity` (`entity_id`) ON DELETE CASCADE ON UPDATE NO ACTION,
   FOREIGN KEY (`service_provider_id`) REFERENCES `service_provider` (`service_provider_id`) ON DELETE CASCADE ON UPDATE NO ACTION
